backend/requirements.txt
# FastAPI and core web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database ORM and migration
sqlalchemy>=2.0.0
alembic>=1.12.0
pymysql>=1.1.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data validation and settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Email and additional utilities
email-validator>=2.0.0

# CORS and middleware
python-multipart>=0.0.6

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Production server (optional)
gunicorn>=21.0.0